requirements.txt
# Runtime dependencies only. Keep dev/test tooling in requirements-dev.txt.
alembic==1.14.1
email-validator==2.2.0
fastapi==0.115.8
pg8000==1.31.5
pydantic==2.10.6
pydantic-settings==2.13.1
PyJWT==2.10.1
python-dotenv==1.0.1
slowapi==0.1.9
SQLAlchemy==2.0.38
uvicorn==0.34.0
